refactor(core): isolate app process quiescence
Move active and superseded process discovery, termination, and regressions into a focused lifecycle submodule so the update fix stays below the repository maintainability limit without changing behavior.
