+++ crates/surge-core/src/update/manager/lifecycle.rs
@@ -1,6 +1,4 @@
 use std::path::Path;
-#[cfg(target_os = "linux")]
-use std::path::PathBuf;
 use std::time::Duration;
 
 use tracing::{debug, info, warn};
@@ -18,6 +16,12 @@ use crate::update::status::{
 const SUPERVISOR_RESTART_CONFIRM_TIMEOUT: Duration = Duration::from_secs(5);
 const SUPERVISOR_RESTART_MAX_ATTEMPTS: u32 = 2;
 const SUPERVISOR_RESTART_RETRY_DELAY: Duration = Duration::from_millis(500);
+
+mod process_quiescence;
+
+pub(super) use self::process_quiescence::{
+    terminate_active_app_processes_before_swap, terminate_superseded_app_processes,
+};
 
 #[derive(Debug, Clone)]
 pub(super) enum SupervisorRestartOutcome {
@@ -290,239 +294,6 @@ fn restart_supervisor_after_update_with_config(
     SupervisorRestartOutcome::PendingRestart { reason, failure_phase }
 }
 
-pub(super) fn terminate_superseded_app_processes(
-    install_dir: &Path,
-    active_app_dir: &Path,
-    main_exe: &str,
-) -> Result<usize> {
-    terminate_superseded_app_processes_except(install_dir, active_app_dir, main_exe, current_pid())
-}
-
-pub(super) fn terminate_active_app_processes_before_swap(active_app_dir: &Path, main_exe: &str) -> Result<usize> {
-    terminate_active_app_processes_except(active_app_dir, main_exe, current_pid())
-}
-
-#[cfg(unix)]
-fn terminate_superseded_app_processes_except(
-    install_dir: &Path,
-    active_app_dir: &Path,
-    main_exe: &str,
-    protected_pid: u32,
-) -> Result<usize> {
-    terminate_matching_app_processes(main_exe, protected_pid, "superseded", |exe| {
-        is_superseded_app_exe(install_dir, active_app_dir, main_exe, exe)
-    })
-}
-
-#[cfg(unix)]
-fn terminate_active_app_processes_except(active_app_dir: &Path, main_exe: &str, protected_pid: u32) -> Result<usize> {
-    let main_exe = main_exe.trim();
-    if main_exe.is_empty() {
-        return Ok(0);
-    }
-
-    let active_exe = match std::fs::canonicalize(active_app_dir.join(main_exe)) {
-        Ok(active_exe) => active_exe,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(0),
-        Err(e) => {
-            return Err(SurgeError::Platform(format!(
-                "Failed to resolve active application executable before swap: {e}"
-            )));
-        }
-    };
-    terminate_matching_app_processes(main_exe, protected_pid, "active", |exe| {
-        is_active_app_exe(&active_exe, exe)
-    })
-}
-
-#[cfg(unix)]
-fn terminate_matching_app_processes<F>(
-    main_exe: &str,
-    protected_pid: u32,
-    process_scope: &'static str,
-    matches_exe: F,
-) -> Result<usize>
-where
-    F: Fn(&Path) -> bool,
-{
-    use nix::errno::Errno;
-    use nix::sys::signal::Signal;
-
-    let main_exe = main_exe.trim();
-    if main_exe.is_empty() {
-        return Ok(0);
-    }
-
-    let pids = app_process_pids(protected_pid, &matches_exe)?;
-    if pids.is_empty() {
-        return Ok(0);
-    }
-
-    for pid in &pids {
-        if let Err(e) = signal_pid(*pid, Signal::SIGTERM) {
-            warn!(pid, error = %e, process_scope, "Failed to request app process termination");
-        }
-    }
-
-    if wait_until_app_processes_exit(protected_pid, &matches_exe, Duration::from_secs(5))? {
-        info!(count = pids.len(), process_scope, "Terminated app processes");
-        return Ok(pids.len());
-    }
-
-    let remaining = app_process_pids(protected_pid, &matches_exe)?;
-    for pid in &remaining {
-        match signal_pid(*pid, Signal::SIGKILL) {
-            Ok(()) | Err(Errno::ESRCH) => {}
-            Err(e) => {
-                warn!(pid, error = %e, process_scope, "Failed to force-kill app process");
-            }
-        }
-    }
-
-    if wait_until_app_processes_exit(protected_pid, &matches_exe, Duration::from_secs(2))? {
-        info!(
-            count = pids.len(),
-            forced = remaining.len(),
-            process_scope,
-            "Force-killed app processes"
-        );
-        return Ok(pids.len());
-    }
-
-    Err(SurgeError::Platform(format!(
-        "Timed out waiting for {process_scope} '{main_exe}' processes to exit"
-    )))
-}
-
-#[cfg(unix)]
-fn signal_pid(pid: u32, signal: nix::sys::signal::Signal) -> std::result::Result<(), nix::errno::Errno> {
-    use nix::sys::signal::kill;
-    use nix::unistd::Pid;
-
-    let Ok(raw_pid) = i32::try_from(pid) else {
-        return Ok(());
-    };
-    kill(Pid::from_raw(raw_pid), signal)
-}
-
-#[cfg(not(unix))]
-fn terminate_superseded_app_processes_except(
-    _install_dir: &Path,
-    _active_app_dir: &Path,
-    _main_exe: &str,
-    _protected_pid: u32,
-) -> Result<usize> {
-    Ok(0)
-}
-
-#[cfg(not(unix))]
-fn terminate_active_app_processes_except(
-    _active_app_dir: &Path,
-    _main_exe: &str,
-    _protected_pid: u32,
-) -> Result<usize> {
-    Ok(0)
-}
-
-#[cfg(unix)]
-fn wait_until_app_processes_exit<F>(protected_pid: u32, matches_exe: &F, timeout: Duration) -> Result<bool>
-where
-    F: Fn(&Path) -> bool,
-{
-    let deadline = std::time::Instant::now() + timeout;
-    loop {
-        if app_process_pids(protected_pid, matches_exe)?.is_empty() {
-            return Ok(true);
-        }
-        if std::time::Instant::now() >= deadline {
-            return Ok(false);
-        }
-        std::thread::sleep(Duration::from_millis(100));
-    }
-}
-
-#[cfg(target_os = "linux")]
-fn app_process_pids<F>(protected_pid: u32, matches_exe: &F) -> Result<Vec<u32>>
-where
-    F: Fn(&Path) -> bool,
-{
-    let entries = std::fs::read_dir("/proc")
-        .map_err(|e| SurgeError::Platform(format!("Failed to enumerate processes from /proc: {e}")))?;
-
-    Ok(entries
-        .filter_map(std::result::Result::ok)
-        .filter_map(|entry| entry.file_name().to_string_lossy().parse::<u32>().ok())
-        .filter(|pid| *pid != protected_pid)
-        .filter(|pid| {
-            std::fs::read_link(format!("/proc/{pid}/exe"))
-                .map(normalize_proc_exe_path)
-                .is_ok_and(|exe| matches_exe(&exe))
-        })
-        .collect())
-}
-
-#[cfg(target_os = "macos")]
-fn app_process_pids<F>(protected_pid: u32, matches_exe: &F) -> Result<Vec<u32>>
-where
-    F: Fn(&Path) -> bool,
-{
-    use sysinfo::{ProcessesToUpdate, System};
-
-    let mut system = System::new();
-    let _ = system.refresh_processes(ProcessesToUpdate::All, true);
-    Ok(system
-        .processes()
-        .iter()
-        .filter_map(|(pid, process)| {
-            let pid = pid.as_u32();
-            (pid != protected_pid && process.exe().is_some_and(matches_exe)).then_some(pid)
-        })
-        .collect())
-}
-
-#[cfg(all(unix, not(any(target_os = "linux", target_os = "macos"))))]
-fn app_process_pids<F>(_protected_pid: u32, _matches_exe: &F) -> Result<Vec<u32>>
-where
-    F: Fn(&Path) -> bool,
-{
-    Err(SurgeError::Platform(
-        "Active application process discovery is unsupported on this Unix platform".to_string(),
-    ))
-}
-
-#[cfg(target_os = "linux")]
-fn normalize_proc_exe_path(path: PathBuf) -> PathBuf {
-    let normalized = {
-        let path_text = path.to_string_lossy();
-        path_text.strip_suffix(" (deleted)").map(PathBuf::from)
-    };
-    normalized.unwrap_or(path)
-}
-
-#[cfg(unix)]
-fn is_active_app_exe(active_exe: &Path, exe: &Path) -> bool {
-    exe == active_exe
-}
-
-#[cfg(unix)]
-fn is_superseded_app_exe(install_dir: &Path, active_app_dir: &Path, main_exe: &str, exe: &Path) -> bool {
-    if exe == active_app_dir.join(main_exe) || exe.file_name().and_then(|name| name.to_str()) != Some(main_exe) {
-        return false;
-    }
-
-    let Ok(relative) = exe.strip_prefix(install_dir) else {
-        return false;
-    };
-
-    let mut components = relative.components();
-    let Some(std::path::Component::Normal(first)) = components.next() else {
-        return false;
-    };
-    let first = first.to_string_lossy();
-
-    first == ".surge-app-prev" || first.starts_with("app-") || components.next().is_none()
-}
-
 fn supervisor_watch_args(
     supervisor_id: &str,
     install_dir: &Path,
@@ -646,98 +417,6 @@ mod tests {
         assert_eq!(
             attempts, 2,
             "restart should spawn the supervisor exactly twice (one retry)"
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn superseded_app_exe_detection_matches_retained_directories_only() {
-        let install_dir = Path::new("/opt/demo");
-        let active_app_dir = install_dir.join("app");
-
-        assert!(is_superseded_app_exe(
-            install_dir,
-            &active_app_dir,
-            "demo",
-            Path::new("/opt/demo/app-1.0.0/demo")
-        ));
-        assert!(is_superseded_app_exe(
-            install_dir,
-            &active_app_dir,
-            "demo",
-            Path::new("/opt/demo/.surge-app-prev/demo")
-        ));
-        assert!(is_superseded_app_exe(
-            install_dir,
-            &active_app_dir,
-            "demo",
-            Path::new("/opt/demo/demo")
-        ));
-        assert!(!is_superseded_app_exe(
-            install_dir,
-            &active_app_dir,
-            "demo",
-            Path::new("/opt/demo/app/demo")
-        ));
-        assert!(!is_superseded_app_exe(
-            install_dir,
-            &active_app_dir,
-            "demo",
-            Path::new("/opt/demo/app-1.0.0/other")
-        ));
-        assert!(!is_superseded_app_exe(
-            install_dir,
-            &active_app_dir,
-            "demo",
-            Path::new("/srv/other/app-1.0.0/demo")
-        ));
-    }
-
-    #[cfg(any(target_os = "linux", target_os = "macos"))]
-    #[test]
-    fn active_app_process_is_terminated_before_swap() {
-        use std::os::unix::fs::{PermissionsExt, symlink};
-        use std::process::Command;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let active_app_dir = tmp.path().join("app");
-        std::fs::create_dir_all(&active_app_dir).unwrap();
-        let app_path = active_app_dir.join("demo");
-        std::fs::copy("/bin/sleep", &app_path).unwrap();
-        let mut permissions = std::fs::metadata(&app_path).unwrap().permissions();
-        permissions.set_mode(0o755);
-        std::fs::set_permissions(&app_path, permissions).unwrap();
-        let linked_active_app_dir = tmp.path().join("linked-app");
-        symlink(&active_app_dir, &linked_active_app_dir).unwrap();
-        let resolved_app_path = std::fs::canonicalize(&app_path).unwrap();
-
-        let mut child = Command::new(&app_path).arg("30").spawn().unwrap();
-        let child_pid = child.id();
-        let deadline = std::time::Instant::now() + Duration::from_secs(1);
-        while !app_process_pids(u32::MAX, &|exe| is_active_app_exe(&resolved_app_path, exe))
-            .unwrap()
-            .contains(&child_pid)
-        {
-            assert!(
-                std::time::Instant::now() < deadline,
-                "test child did not enter the active app path"
-            );
-            std::thread::sleep(Duration::from_millis(10));
-        }
-
-        let terminated = terminate_active_app_processes_except(&linked_active_app_dir, "demo", u32::MAX).unwrap();
-        let status = child.wait().unwrap();
-
-        assert_eq!(terminated, 1);
-        assert!(!status.success());
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn proc_exe_deleted_suffix_is_ignored_for_matching() {
-        assert_eq!(
-            normalize_proc_exe_path(PathBuf::from("/opt/demo/app-1.0.0/demo (deleted)")),
-            PathBuf::from("/opt/demo/app-1.0.0/demo")
         );
     }
 }
